+++ js/mootools/components/Tooltip.js
@@ -162,7 +162,7 @@ Toolkit.Tooltip = new Class({
                 this.position(this.cache[content], title);
             } else {
                 if (options.showLoading) {
-                    this.position(options.loadingMessage);
+                    this.position(Toolkit.messages.loading);
                 }
 
                 this.requestData(content);