+++ changelog.md
@@ -6,8 +6,10 @@ Older versions can be found in the documentation changelogs.
 This update includes documentation for all components.
 
 * Added `isRetina` to the `Toolkit` flags.
+* Removed global `loadingMessage` and `errorMessage` options. Will all use the messages found in `Toolkit`.
 * Blackout
     * Added a `shown` argument to the `show` event
+    * Added `showLoader` and `hideLoader` events
     * Renamed `loaderMessage` option to `loadingMessage`
 * Button
     * Added outline none to `:focus` styles